Read fractional mAh figures correctly
batterystats formats charge by magnitude: two decimals below 10 mAh, one
below 100, whole numbers above. The mAh pattern matched integers only,
which did not fail on the decimal forms — it skipped the whole number and
matched the digits after the point.

  "137 mAh"  -> 137   correct by luck
  "93.4 mAh" -> 4
  "5.56 mAh" -> 56    ten times too large

So the drain card contradicted the discharge total sitting beside it: a
136 mAh cycle showed 128 mAh screen-on against 56 mAh screen-off, split
70/30, with a screen-off rate of 2349 mA.

Accept the fractional part, and carry these through as Double rather than
rounding. The rounding would still have turned the 7.56 mAh a cycle
reports in its first minutes into 8.

The fixture that missed this is one device's dump that happens to be all
whole numbers. Cover all three magnitudes instead, plus a signed value.

+++ app/src/main/java/com/syed/wattson/data/model/BatteryModels.kt
@@ -9,12 +9,12 @@ package com.syed.wattson.data.model
  * presented as a total must come from here.
  */
 data class MeasuredDischarge(
-    val totalMah: Int,
-    val screenOnMah: Int?,
-    val screenOffMah: Int?,
-    val screenDozeMah: Int?,
-    val lightDozeMah: Int?,
-    val deepDozeMah: Int?,
+    val totalMah: Double,
+    val screenOnMah: Double?,
+    val screenOffMah: Double?,
+    val screenDozeMah: Double?,
+    val lightDozeMah: Double?,
+    val deepDozeMah: Double?,
 )
 
 /** Everything parsed out of `dumpsys batterystats --charged`. */
@@ -25,7 +25,12 @@ data class BatteryStats(
     val screenOffMs: Long,
     val screenOnCount: Int,
     val totalRunTimeMs: Long,
-    val dischargeMah: Int?,
+    /**
+     * Charge taken out of the cell this cycle. Kept fractional: batterystats prints these
+     * to two decimals below 10 mAh and one below 100, and rounding to a whole number
+     * throws away most of the figure for the first minutes of a cycle.
+     */
+    val dischargeMah: Double?,
     val designCapacityMah: Int?,
     val powerByState: PowerByState,
     val measured: MeasuredDischarge? = null,
